+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changes in this section will be included in the next release.
 #### Added
 
 - Allowed the ability to persist project selection with `select project foo`. If set via this method, `--project` will no longer be a required flag for other commands.
+- Added `batch wait` command that will continuously poll (default: 30s) until a batch reaches a final state (success, failure, cancelled, etc)
 
 #### Changed